[test] Split onboarding and step objects tests
